ai-service/main.py
import openai
from fastapi import FastAPI
import gemini
import plotly
import markdown

app = FastAPI()

@app.get("/")
def read_root():
    return {"Hello": "World"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
