ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.123.0",
    "markdown>=3.10",
    "openai>=2.8.1",
    "plotly>=6.5.0",
    "python-gemini-api>=2.4.12",
    "uvicorn>=0.25.0"
]
